prepare repository to be cloned in cmder config
This repo is intended to contain files to be placed at `cmder/config/`, which contains lots of other files that are not part of this repo that need to be ignored. One of them is a README.md, so we rename this repo's README with a different extension that works on Github.